pyproject.toml
[project]
name = "genie-workbench"
version = "0.1.0"
description = "Unified Databricks Genie Space management platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "databricks-sdk>=0.58.0",
    "pydantic>=2.0",
    "python-dotenv",
    "mlflow>=3.6.0",
    "databricks-agents>=1.4.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "openai>=1.0",
    "requests>=2.31.0",
    "streamlit>=1.40.0",
    "asyncpg>=0.29",
    "psycopg2-binary",
    "mcp>=1.26.0",
    "pyyaml>=6.0",
]

[project.scripts]
start-server = "backend.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]
